Added templates and fixed redirect after login based on role

+++ src/main/java/com/helmes/assignment/entity/repositories/MyUserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MyUserRepository extends JpaRepository<MyUser, Integer> {
+public interface MyUserRepository extends JpaRepository<MyUser, Long> {
     Optional<MyUser> findByUsername(String username);
 }
